+++ barney/volume/MCAccelerator.h
@@ -303,7 +303,7 @@ namespace barney {
         MCVolumeAccel<SFSampler>::DD::addVars(params,0);
         OWLGeomType gt = owlGeomTypeCreate
           (dg->owl,OWL_GEOM_USER,
-           sizeof(MCVolumeAccel<SFSampler>::DD),params.data(),params.size());
+           sizeof(MCVolumeAccel<SFSampler>::DD),params.data(),(int)params.size());
         OWLModule module = owlModuleCreate(dg->owl,ptxCode);
         const std::string boundsProg = typeString+"_Bounds";
         const std::string isProg = typeString+"_Isec";
